+++ src/OpenTelemetry.Instrumentation.EntityFrameworkCore/CHANGELOG.md
@@ -7,6 +7,7 @@
 Released 2024-Feb-07
 
 * **Breaking Change**: Stop emitting `db.statement_type` attribute.
+  This attribute never was part of the [semantic convention](https://github.com/open-telemetry/semantic-conventions/blob/v1.24.0/docs/database/database-spans.md#call-level-attributes).
   ([#1559](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1559))
 
 ## 1.0.0-beta.9